Makefile: Disable build-id so we can build with Bionic's gcc
Fix building the tests using the gcc cross-compilers shipped
with Ubuntu Bionic. These compilers now default to outputting
a build-id; our linker script is not expecting that, so the
linker defaults to putting it somewhere that turns out to
conflict with other things and results in a "not enough room
for program headers" error from the linker.

We could make the linker script just discard the build-id,
but this apparently results in it complaining that it has
done so:
 https://stackoverflow.com/questions/15316384/do-not-pass-build-id-to-linker-from-gcc

So instead we just tell the linker not to create the build-id
section in the first place.

Since this means the linker command line options are now
pretty long and repetitive, we factor them out into some
make variables.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>